- Added support for total row count - Updated offset to support long - Added support for with() syntax when selecting

+++ src/main/java/be/nabu/libs/services/jdbc/DefaultDialect.java
@@ -10,7 +10,7 @@ public class DefaultDialect implements SQLDialect {
 		return sql;
 	}
 	@Override
-	public String limit(String sql, Integer offset, Integer limit) {
+	public String limit(String sql, Long offset, Integer limit) {
 		return sql;
 	}
 	@Override
